+++ refiner/tests/test_configurations.py
@@ -1,56 +1,19 @@
-from contextlib import asynccontextmanager
 from unittest.mock import AsyncMock
 from uuid import UUID
 
 import pytest
-import pytest_asyncio
-from httpx import ASGITransport, AsyncClient
 
-from app.api.auth.middleware import get_logged_in_user
 from app.api.v1.configurations import GetConfigurationsResponse
 from app.db.conditions.model import DbCondition
 from app.db.configurations.db import DbTotalConditionCodeCount
 from app.db.user.model import DbUser
-from app.main import app
-
-TEST_SESSION_TOKEN = "test-token"
-
-
-class MockDB:
-    @asynccontextmanager
-    async def get_connection(self):
-        mock_conn = AsyncMock()
-        yield mock_conn
-
-
-@pytest_asyncio.fixture
-async def authed_client(mock_logged_in_user, mock_db_functions):
-    transport = ASGITransport(app=app)
-    async with AsyncClient(transport=transport, base_url="http://testserver") as client:
-        client.cookies.update({"refiner-session": TEST_SESSION_TOKEN})
-        yield client
-
-
-@pytest.fixture(autouse=True)
-def mock_logged_in_user():
-    """Mock the logged-in user dependency"""
-
-    def mock_user():
-        return {
-            "id": "5deb43c2-6a82-4052-9918-616e01d255c7",
-            "username": "tester",
-            "email": "tester@test.com",
-            "jurisdiction_id": "JD-1",
-        }
-
-    app.dependency_overrides[get_logged_in_user] = mock_user
-    yield
-    app.dependency_overrides.pop(get_logged_in_user, None)
 
 
 @pytest.fixture(autouse=True)
 def mock_db_functions(monkeypatch):
-    """Mock DB functions directly to avoid connections/cursors."""
+    """
+    Mock return values of the `_db` functions called by the routes.
+    """
     # Mock get_user_by_id_db
     user_mock = DbUser(
         id="5deb43c2-6a82-4052-9918-616e01d255c7",
@@ -84,7 +47,7 @@ def mock_db_functions(monkeypatch):
         AsyncMock(return_value=config_by_id_mock),
     )
 
-    # Mock is config valid to insert
+    # Mock is_config_valid_to_insert_db
     monkeypatch.setattr(
         "app.api.v1.configurations.is_config_valid_to_insert_db",
         AsyncMock(return_value=True),
